Added chunks and multithreading
At least, the scaffolding for proper chunks and multithreading is in place. The World class generates a list of Chunk objects during initialization, and each Chunk is processed in a set order using multithreading to avoid race conditions and allow for faster processing of Pixels.

I did some testing and I'm *pretty sure* it's working as intended, but I won't really be able to test it until I add more features like dirty rectangles and Chunk sleeping/waking.